hromadne operacie

+++ locallib.php
@@ -46,11 +46,13 @@ function gallery_process_drafts($context, $gallery) {
     foreach($files as $file) {
         if(!$file->is_valid_image()) {
             $packer = get_file_packer($file->get_mimetype());
-            $file->extract_to_storage($packer, $context->id, 'mod_gallery', 'unpacktemp', $i, '/');
-            $unpackedFiles = $fs->get_area_files($context->id, 'mod_gallery', 'unpacktemp', $i,'filename ASC');
-            $preloaded_images = array_merge($preloaded_images, $unpackedFiles);
-            $file->delete();
-            $i++;
+            if($packer) {
+                $file->extract_to_storage($packer, $context->id, 'mod_gallery', 'unpacktemp', $i, '/');
+                $unpackedFiles = $fs->get_area_files($context->id, 'mod_gallery', 'unpacktemp', $i,'filename ASC');
+                $preloaded_images = array_merge($preloaded_images, $unpackedFiles);
+                $file->delete();
+                $i++;
+            }
         } 
             $preloaded_images[] = $file;
     }
